Remove unused utility functions from project module
In this commit, a number of utility functions that were not being used in the 'project' module of the 'ppath' project were removed. These functions included 'findfile', 'findup', 'flatten' and some others. Their removal streamlines the codebase and ensures it only includes code that is necessary for the functioning of the project.