Move PY_IGNORE_IMPORTMISMATCH environmental variable assignment.
The PY_IGNORE_IMPORTMISMATCH environmental variable was declared in src/_backup/__init__.py file which was not optimally placed. This variable influences the functioning of the entire application, hence should be placed at the root level. So this commit moves the assignment of PY_IGNORE_IMPORTMISMATCH from the src/_backup/__init__.py to src/nodeps/__init__.py. This change is in favour of improving the variables scope and making the code more sustainable.