Update code references and running environment
Made adjustments in the code referencing __file__ instead of specific dependencies for better generalization. Changed the python environment from slim-buster to slim in tests.dockerfile for better compatibility. Updated the function `which` in functions.py to check only for executables. Adjusted the `chdir` logic in functions.py to restrict code paths exclusively to MacOS. Removed an unnecessary docker dependency in pyproject.toml. Altered the pip-tools flag from --all-extras to --no-strip-extras to prevent removal of optional dependencies.