Update GitHub Actions workflow to add OS info and specify test path
The workflow configuration in main.yml is updated to include the operating system in the test job name. This change adds clarity when reading the CI reports to understand on which OS the tests are running. Besides, the 'run' command for pytest was modified to only run tests found in the 'src/nodeps/modules/gh.py' file as focusing the tests on this specific module is currently necessary.