Skip commented out pytest run command
The commented out pytest run command in the Github Action workflow file was skipped to clean up the code. The specific test file `test_cli_project_p.py` with the `test_build_path` test is now explicitly run instead as it is more relevant to the recent changes and needs to be specifically tested.