Fix permission checks and comment out 'elevate()'
Commented out the evaluation of whether the system is not MACOS and if the path checked has write permission, as well as the subsequent 'elevate()' function in functions.py. This is a temporary solution to avoid unnecessary privilege escalation that could potentially compromise system security. The decision to comment this out instead of removing is to potentially revisit for debugging and optimization based on future requirements.