Corrected default passwd handling and updated tests in main.yml
Two changes were made within this commit. Firstly, in the 'path.py' file, the passwd handling has been updated. If the passwd was neither a string nor provided, it used to be set to 'Passwd(passwd)', now it falls back to the USER defined in the environment instead of raising an error. This change provides a more robust default behaviour.

Secondly, in the 'main.yml' file, the pytest command has been adjusted. Previously the 'path.py' file was called directly, but now it calls the function 'test_build_path' within 'test_cli_project_p.py'. This change allows for more specific testing.