Update pytest command and refine passwd handling in main.yml
This commit updates the pytest command in the 'main.yml' to run the 'test_build_path' function from 'test_cli_project_p.py', instead of directly calling the 'path.py'. This allows for more specific testing. In addition, the passwd handling has been refined. If the passwd is neither a string nor provided, it will now default to the USER defined in the environment, instead of raising an error, providing a more robust default behaviour.