Add -y flag to pip uninstall command in GitHub workflow
The -y flag was added to the 'pip uninstall nodeps' command under Install dependencies in the .github/workflows/main.yml file. This change ensures the uninterrupted flow of the workflow as it avoids manually entering 'yes' as confirmation for uninstalling the package. It keeps the working environment clean by removing unnecessary packages.