Refactor DOCKER checking in tests and fixtures
Changes involve modifications to how DOCKER environment variable is checked in several test functions and fixtures. These changes were made for code correction and consistency. Previously, the code falsely considered the test to be in a docker environment when the DOCKER variable was False. The changes bring it back to its intended functionality: to check if the tests are running in a Docker environment. Test function and fixture descriptions were also updated to correctly reflect this behaviour.